StreamLeg => Stream.StepLeg minor comment typos
